backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
google-genai>=0.3.0
google-cloud-storage>=2.10.0
google-cloud-firestore>=2.13.0
opencv-python>=4.8.0
moviepy>=1.0.3
python-dotenv>=1.0.0
gunicorn>=21.2.0
requests>=2.31.0
pillow>=10.0.0
werkzeug>=3.0.0
